Building: split Building.cpp (2896 lines) into 6 focused TUs
| File                          | Lines | Contents                                                                                               |
|-------------------------------|------:|--------------------------------------------------------------------------------------------------------|
| src/building/Misc.cpp         |   539 | kill, unit/ressource management, exits, level, flag stats, eatOnce, happiness, conversion, integrity, checkSum |
| src/building/Step.cpp         |   525 | step, subscribeToBringRessourcesStep, subscribeForFlagingStep, subscribeUnitForInside                  |
| src/building/Lifecycle.cpp    |   527 | constructors, destructor, freeGradients, load/save, loadCrossRef/saveCrossRef                          |
| src/building/Update.cpp       |   470 | building-site / units / general updates, upgrade-area forbidden zones, isHardSpaceForBuildingSite      |
| src/building/Construction.cpp |   415 | resource needs/wishes, launch/cancel construction & delete, updateCallLists, updateConstructionState   |
| src/building/TypeSteps.cpp    |   382 | swarmStep, turretStep, clearingFlagStep                                                                |

Pure extraction: every line is a byte-exact copy of the original.
Each TU duplicates the original 38-line prelude (copyright + includes).
The mid-file `#include "GameGUI.h"` (was line 1333) travels with Update.cpp.
Building.h is unchanged. SConstruct adds `#src` to CPPPATH so the moved
files can still find headers that live in src/.
