+++ src/building/TypeSteps.cpp
@@ -1,0 +1,420 @@
+/*
+  Copyright (C) 2001-2004 Stephane Magnenat & Luc-Olivier de Charrière
+  for any question or comment contact us at <stephane at magnenat dot net> or <NuageBleu at gmail dot com>
+
+  This program is free software; you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation; either version 3 of the License, or
+  (at your option) any later version.
+
+  This program is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with this program; if not, write to the Free Software
+  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+*/
+
+#include <list>
+#include <math.h>
+#include <Stream.h>
+#include <stdlib.h>
+#include <algorithm>
+#include <climits>
+
+#include "Building.h"
+#include "BuildingType.h"
+#include "Game.h"
+#include "GlobalContainer.h"
+#include "LogFileManager.h"
+#include "Team.h"
+#include "Unit.h"
+#include "Utilities.h"
+#include "Order.h"
+#include "Bullet.h"
+#include "Integrity.h"
+
+void Building::swarmStep(void)
+{
+	// increase HP
+	if (hp<type->hpMax)
+		hp++;
+	assert(NB_UNIT_TYPE==3);
+	if ((ressources[CORN]>=type->ressourceForOneUnit)&&(ratio[0]|ratio[1]|ratio[2]))
+		productionTimeout--;
+
+	if (productionTimeout<0)
+	{
+		// We find the kind of unit we have to create:
+		Sint32 fProportion;
+		Sint32 fMinProportion=0x7FFFFFFF;
+		int minType=-1;
+		for (int i=0; i<NB_UNIT_TYPE; i++)
+			if (ratio[i]!=0)
+			{
+				fProportion=(percentUsed[i]<<16)/ratio[i];
+				if (fProportion<=fMinProportion)
+				{
+					fMinProportion=fProportion;
+					minType=i;
+				}
+			}
+
+		if (minType==-1)
+			minType=0;
+		assert(minType>=0);
+		assert(minType<NB_UNIT_TYPE);
+		if (minType<0 || minType>=NB_UNIT_TYPE)
+			minType=0;
+
+		// We get the unit UnitType:
+		int posX, posY, dx, dy;
+		UnitType *ut=owner->race.getUnitType(minType, 0);
+
+		// Is there a place to exit ?
+		bool exitFound;
+		if (ut->performance[FLY])
+			exitFound=findAirExit(&posX, &posY, &dx, &dy);
+		else
+			exitFound=findGroundExit(&posX, &posY, &dx, &dy, ut->performance[SWIM]);
+		if (exitFound)
+		{
+			Unit * u=owner->game->addUnit(posX, posY, owner->teamNumber, minType, 0, 0, dx, dy);
+			if (u)
+			{
+				ressources[CORN]-=type->ressourceForOneUnit;
+				updateCallLists();
+
+				u->activity=Unit::ACT_RANDOM;
+				u->displacement=Unit::DIS_RANDOM;
+				u->movement=Unit::MOV_EXITING_BUILDING;
+				u->speed=u->performance[u->action];
+
+				productionTimeout=type->unitProductionTime;
+
+				// We update percentUsed[]
+				percentUsed[minType]++;
+
+				bool allDone=true;
+				for (int i=0; i<NB_UNIT_TYPE; i++)
+					if (percentUsed[i]<ratio[i])
+						allDone=false;
+
+				if (allDone)
+					for (int i=0; i<NB_UNIT_TYPE; i++)
+						percentUsed[i]=0;
+			}
+			else if (verbose)
+				printf("WARNING, no more UNIT ID free for team %d\n", owner->teamNumber);
+		}
+	}
+}
+
+
+void Building::turretStep(Uint32 stepCounter)
+{
+	// create bullet from stones in stock
+	if (ressources[STONE]>0 && (bullets<=(type->maxBullets-type->multiplierStoneToBullets)))
+	{
+		ressources[STONE]--;
+		bullets += type->multiplierStoneToBullets;
+
+		// we need to be stone-feeded
+		updateCallLists();
+	}
+
+	// compute cooldown
+	if (shootingCooldown > 0)
+	{
+		shootingCooldown -= type->shootRythme;
+		return;
+	}
+
+	// if we have no bullet, don't try to shoot
+	if (bullets <= 0)
+		return;
+
+	//for some reason, any turret that is not 2x2 makes no sense at all to the game
+	assert(type->width ==2);
+	assert(type->height==2);
+
+	int range = type->shootingRange;
+	shootingStep = (shootingStep+1)&0x7;
+
+	Uint32 enemies = owner->enemies;
+	Map *map = owner->map;
+	assert(map);
+
+	// the type of target we have found
+	enum TargetType
+	{
+		TARGETTYPE_NONE,
+		TARGETTYPE_BUILDING,
+		TARGETTYPE_WORKER,
+		TARGETTYPE_WARRIOR,
+		TARGETTYPE_EXPLORER,
+	};
+	// The type of the best target we have found up to now
+	TargetType targetFound = TARGETTYPE_NONE;
+	// The score of the best target we have found up to now
+	int bestScore = INT_MIN;
+	// The number of ticks before the unit may move away
+	int bestTicks = 0;
+	// The position of the best target we have found up to now
+	int bestTargetX = 0, bestTargetY=0;
+
+	for (int i=0; i<=range ; i++)
+	{
+		// The number of ticks before the bullet hits the target at range "i".
+		int ticksToHit = ((i << 5) + ((type->width) << 4)) / (type->shootSpeed>>8);
+		for (int j=0; j<=i ; j++)
+		{
+			for (int k=0; k<8; k++)
+			{
+				int targetX, targetY;
+				switch (k)
+				{
+					case 0:
+					targetX=posX-j;
+					targetY=posY-i;
+					break;
+					case 1:
+					targetX=posX+j+1;
+					targetY=posY-i;
+					break;
+					case 2:
+					targetX=posX-j;
+					targetY=posY+i+1;
+					break;
+					case 3:
+					targetX=posX+j+1;
+					targetY=posY+i+1;
+					break;
+					case 4:
+					targetX=posX-i;
+					targetY=posY-j;
+					break;
+					case 5:
+					targetX=posX+i+1;
+					targetY=posY-j;
+					break;
+					case 6:
+					targetX=posX-i;
+					targetY=posY+j+1;
+					break;
+					case 7:
+					targetX=posX+i+1;
+					targetY=posY+j+1;
+					break;
+					default:
+					assert(false);
+					targetX=0;
+					targetY=0;
+					break;
+				}
+				int targetGUID = map->getGroundUnit(targetX, targetY);
+				int airTargetGUID = map->getAirUnit(targetX, targetY);
+				if (targetGUID != NOGUID)
+				{
+					Sint32 otherTeam = Unit::GIDtoTeam(targetGUID);
+					Sint32 targetID = Unit::GIDtoID(targetGUID);
+					Uint32 otherTeamMask = 1<<otherTeam;
+					if (enemies & otherTeamMask)
+					{
+						Unit *testUnit = owner->game->teams[otherTeam]->myUnits[targetID];
+						if ((owner->sharedVisionExchange & otherTeamMask) == 0)
+						{
+							int targetTicks = (256 - testUnit->delta) / testUnit->speed;
+							// skip this unit if it will move away too soon.
+							if (targetTicks <= ticksToHit)
+								continue;
+							// shoot warrior first, then workers if no warrior
+							if (testUnit->typeNum == WARRIOR)
+							{
+								int targetOffense = (testUnit->getRealAttackStrength() * testUnit->performance[ATTACK_SPEED]); // 88 to 1024
+								int targetWeakeness = 0; // 0 to 512
+								if (testUnit->hp > 0)
+								{
+									if (testUnit->hp < type->shootDamage) // hahaha, how mean!
+										targetWeakeness = 512;
+									else
+										targetWeakeness = 256 / testUnit->hp;
+								}
+								int targetProximity = 0; // 0 to 512
+								if (i <= 0)
+									targetProximity = 512;
+								else
+									targetProximity = (256 / i);
+								int targetScore = targetOffense + targetWeakeness + targetProximity;
+								// lower scores are overriden
+								if (targetScore > bestScore)
+								{
+									bestScore = targetScore;
+									bestTicks = targetTicks;
+									bestTargetX = targetX;
+									bestTargetY = targetY;
+									targetFound = TARGETTYPE_WARRIOR;
+								}
+							}
+							else if ((targetFound != TARGETTYPE_WARRIOR) && (testUnit->typeNum == WORKER))
+							{
+								// adjust score for range
+								int targetScore = - testUnit->hp;
+								// lower scores are overriden
+								if (targetScore > bestScore)
+								{
+									bestScore = targetScore;
+									bestTicks = targetTicks;
+									bestTargetX = targetX;
+									bestTargetY = targetY;
+									targetFound = TARGETTYPE_WORKER;
+								}
+							}
+						}
+					}
+				}
+				//explorers are now priority targets as defined later
+
+				if (airTargetGUID != NOGUID)
+				{
+					Sint32 otherTeam = Unit::GIDtoTeam(airTargetGUID);
+					Sint32 targetID = Unit::GIDtoID(airTargetGUID);
+					Uint32 otherTeamMask = 1<<otherTeam;
+					if (enemies & otherTeamMask)
+					{
+						Unit *testUnit = owner->game->teams[otherTeam]->myUnits[targetID];
+						if ((owner->sharedVisionExchange & otherTeamMask) == 0)
+						{
+							int targetTicks = (256 - testUnit->delta) / testUnit->speed;
+							// skip this unit if it will move away too soon.
+							if (targetTicks <= ticksToHit)
+								continue;
+							//Using simple calculation for now (should always shoot ground-attackers first, probably)
+							// adjust score for range
+							int targetScore = - testUnit->hp;
+							// lower scores are overriden
+							if (targetScore > bestScore)
+							{
+								bestScore = targetScore;
+								bestTicks = targetTicks;
+								bestTargetX = targetX;
+								bestTargetY = targetY;
+								targetFound = TARGETTYPE_EXPLORER;
+							}
+						}
+					}
+				}
+
+				// shoot building only if no unit is found
+				if (targetFound == TARGETTYPE_NONE)
+				{
+					Uint16 targetGBID = map->getBuilding(targetX, targetY);
+					if (targetGBID != NOGBID)
+					{
+						Sint32 otherTeam = Building::GIDtoTeam(targetGBID);
+						//int otherID = Building::GIDtoID(targetGBID);
+						Uint32 otherTeamMask = 1<<otherTeam;
+						if (enemies & otherTeamMask)
+						{
+							// adjust score for range
+							int targetScore = - i;
+							if (targetScore > bestScore)
+							{
+								bestScore = targetScore;
+								bestTicks = 256;
+								bestTargetX = targetX;
+								bestTargetY = targetY;
+								targetFound = TARGETTYPE_BUILDING;
+							}
+						}
+					}
+				}
+			}
+		}
+		if (targetFound == TARGETTYPE_EXPLORER)
+			break;//specifying explorers as high priority
+	}
+
+	if (targetFound != TARGETTYPE_NONE)
+	{
+		shootingStep = 0;
+
+		//printf("%d found target found: (%d, %d) \n", gid, targetX, targetY);
+		Sector *s=owner->map->getSector(getMidX(), getMidY());
+
+		int px, py;
+		px=((posX)<<5)+((type->width)<<4);
+		py=((posY)<<5)+((type->height)<<4);
+
+		int speedX, speedY, ticksLeft;
+
+		// TODO : shall we really uses shootSpeed ?
+		// FIXME : is it correct this way ? Is there a function for this ?
+		int dpx=(bestTargetX*32)+16-4-px; // 4 is the half size of the bullet
+		int dpy=(bestTargetY*32)+16-4-py;
+		//printf("%d insert: dp=(%d, %d).\n", gid, dpx, dpy);
+		if (dpx>(map->getW()<<4))
+			dpx=dpx-(map->getW()<<5);
+		if (dpx<-(map->getW()<<4))
+			dpx=dpx+(map->getW()<<5);
+		if (dpy>(map->getH()<<4))
+			dpy=dpy-(map->getH()<<5);
+		if (dpy<-(map->getH()<<4))
+			dpy=dpy+(map->getH()<<5);
+
+		int mdp;
+
+		assert(dpx);
+		assert(dpy);
+		if (abs(dpx)>abs(dpy)) //we avoid a square root, since all ditances are squares lengthed.
+		{
+			mdp=abs(dpx);
+			speedX=((dpx*type->shootSpeed)/(mdp<<8));
+			speedY=((dpy*type->shootSpeed)/(mdp<<8));
+			assert(speedX!=0);
+			ticksLeft=abs(mdp/speedX);
+		}
+		else
+		{
+			mdp=abs(dpy);
+			speedX=((dpx*type->shootSpeed)/(mdp<<8));
+			speedY=((dpy*type->shootSpeed)/(mdp<<8));
+			assert(speedY!=0);
+			ticksLeft=abs(mdp/speedY);
+		}
+
+		if (ticksLeft < bestTicks)
+		{
+			Bullet *b = new Bullet(px, py, speedX, speedY, ticksLeft, type->shootDamage, bestTargetX, bestTargetY, posX-1, posY-1, type->width+2, type->height+2);
+			s->bullets.push_front(b);
+			bullets--;
+			shootingCooldown = SHOOTING_COOLDOWN_MAX;
+			lastShootStep = stepCounter;
+			lastShootSpeedX = speedX;
+			lastShootSpeedY = speedY;
+		}
+	}
+
+}
+
+
+
+void Building::clearingFlagStep()
+{
+	if (unitsWorking.size()<(unsigned)maxUnitWorking)
+		for (int canSwim=0; canSwim<2; canSwim++)
+			if (localRessourcesCleanTime[canSwim]++>125) // Update every 5[s]
+			{
+				if (!owner->map->updateLocalRessources(this, canSwim))
+				{
+					for (std::list<Unit *>::iterator it=unitsWorking.begin(); it!=unitsWorking.end(); ++it)
+						(*it)->standardRandomActivity();
+					unitsWorking.clear();
+				}
+			}
+}
+
+
+
